Move image-saving code to separate file.
